+++ modules/documentation/docs/changelog.md
@@ -2,6 +2,11 @@
 
 ## Next Release
 
+## 0.1.15-beta.17
+
+- \[server-node\] Filter transfers by definition when getting unsubmitted withdrawals
+- \[server-node\] Improve logging for admin endpoint
+
 ## 0.1.15-beta.16
 
 - \[server-node\] Hold withdrawals as alice on mainnet